Timers Test
The timer now creates correctly a popup window with the option to
disable the timer/event 1min before the specified time. If it is not
disabled within one minute then the actual timer goes off and notifies
the contacts.
Also trying to deal with the scenario when the user manually destroys
the app.