Main Screen and Settings
Changed the main screen and the settings screen according to the second
iteration design.
Merged the Emergency Numbers into the Settings screen.
Added a Back option to the Title bar of Settings screen.
Added a screen title bar in the Settings screen.
Added functionality to edit the Emergency message, accessible via the
Settings. It will also persist through killed app instances.
When the Emergency button is triggered, it will create a notification on
the status bar.